purview/api/py/pyproject.toml
[tool.poetry]
name = "davidkhala.purview"
version = "0.0.0"
description = ""
authors = ["David Liu <david-khala@hotmail.com>"]
readme = "README.md"
packages = [{ include = "davidkhala" }]

[tool.poetry.dependencies]
python = "^3.12"
azure-identity = "*"
azure-purview-catalog = "*"
davidkhala-databricks = { version = "*", extras = ["connect"] }


[tool.poetry.group.dev.dependencies]
davidkhala-utils = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
